'评论'

+++ src/main/java/com/hoodee/community/Controlller/PublishController.java
@@ -1,8 +1,6 @@
 package com.hoodee.community.Controlller;
 
 import com.hoodee.community.dto.QuestionDTO;
-import com.hoodee.community.mapper.UserMapper;
-import com.hoodee.community.mapper.QuestionMapper;
 import com.hoodee.community.model.Question;
 import com.hoodee.community.model.User;
 import com.hoodee.community.service.QuestionService;
@@ -15,8 +13,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.servlet.http.HttpServletRequest;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 
 /**
  * Package: com.hoodee.community.Controlller
@@ -32,7 +28,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,Model model){
+    public String edit(@PathVariable(name = "id") Long id,Model model){
 
         QuestionDTO question = questionService.getByID(id);
         model.addAttribute("title",question.getTitle());
@@ -52,7 +48,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value ="description",required = false) String description,
             @RequestParam(value ="tag",required = false) String tag ,
-            @RequestParam(value ="id",required = false) Integer id ,
+            @RequestParam(value ="id",required = false) Long id ,
             HttpServletRequest request, Model model){
 
         model.addAttribute("title",title);
